Rename processUpload to processSampleUpload
